Adde Airport provider (with very little data.  I will add an en_US provider with much more.
One interesting note:    in addition to the normal airport() class method that returns a simple string,
this provide defines airportObject() that returns a richer Airport object that has code, location and name.
I think this model would be useful for other faker providers. For example  faker.family() could return
a list of people objects each of which has a name, a gender, a role (Father, Mother, Son, Aunt), etc.
It remains to be seen if this is stretching the faker interface too far.